Refactoring of package and organising it
