+++ src/main/java/ch/heigvd/res/chill/domain/ClarFleu/FinDuMonde.java
@@ -1,12 +1,13 @@
-package ch.heigvd.res.chill.domain.wasadigi;
+package ch.heigvd.res.chill.domain.ClarFleu;
 
 import ch.heigvd.res.chill.domain.IProduct;
 
 import java.math.BigDecimal;
 
-public class Kwak implements IProduct {
-    public final static String NAME = "Kwak";
-    public final static BigDecimal PRICE = new BigDecimal(4.2);
+public class FinDuMonde implements IProduct {
+
+    public final static String NAME = "Fin du Monde";
+    public final static BigDecimal PRICE = new BigDecimal(3.0);
 
     @Override
     public String getName() {
